Consolidate IR serializer-field and executor-class builders
StubGeneration and ObjGeneration each carried near-identical copies of:

- Private `Name = "${Tn}Serializer"` IrField creation per type parameter
  on the generated Stub/Obj class.
- Anonymous `ServiceExecutor` implementation class construction
  (primary ctor delegating to Any(), `invoke` override that dispatches
  to the referenced service method with implicit casts on the
  receiver and value argument).

Extract these into `buildSerializerFieldsForTypeParams` and
`buildAnonymousServiceExecutor` in IrUtils.kt and call them from both
transformers. Also name the field/parameter scheme via a shared
`serializerFieldName(Name)` helper so the FIR and IR sides cannot
drift.

No behavior change.
